extract common_recref_adapter
